use vyxal modulo for indexing bc scala modulo is weird

+++ shared/src/vyxal/elements/Elements.scala
@@ -783,7 +783,7 @@ object Elements:
       },
     addPart("Þ⦷", Monad, true) {
       case i: VNum =>
-        RegisterHelpers.index(i.toInt)
+        RegisterHelpers.index(i)
       },
     addPart("Þ⊖", Monad, false) {
       case n: VNum =>
